Split custom JSON writer and parser to separate files
