+++ protobuf/lib/protobuf.dart
@@ -34,7 +34,8 @@ part 'src/protobuf/field_set.dart';
 part 'src/protobuf/field_type.dart';
 part 'src/protobuf/generated_message.dart';
 part 'src/protobuf/generated_service.dart';
-part 'src/protobuf/json.dart';
+part 'src/protobuf/json_reader.dart';
+part 'src/protobuf/json_writer.dart';
 part 'src/protobuf/pb_list.dart';
 part 'src/protobuf/pb_map.dart';
 part 'src/protobuf/protobuf_enum.dart';
